src/main/java/nextstep/subway/auth/ui/interceptor/authorization/TokenSecurityContextPersistenceInterceptor.java
package nextstep.subway.auth.ui.interceptor.authorization;

import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.ObjectMapper;
import nextstep.subway.auth.domain.Authentication;
import nextstep.subway.auth.infrastructure.*;
import nextstep.subway.member.domain.LoginMember;
import org.springframework.web.servlet.HandlerInterceptor;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

public class TokenSecurityContextPersistenceInterceptor implements HandlerInterceptor {
    private JwtTokenProvider jwtTokenProvider;

    public TokenSecurityContextPersistenceInterceptor(JwtTokenProvider jwtTokenProvider) {
        this.jwtTokenProvider = jwtTokenProvider;
    }

    @Override
    public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) throws JsonProcessingException {
        String jwtToken = AuthorizationExtractor.extract(request, AuthorizationType.BEARER);
        boolean validated = jwtTokenProvider.validateToken(jwtToken);

        if (validated) {
            LoginMember loginMember = new ObjectMapper().readValue(jwtTokenProvider.getPayload(jwtToken), LoginMember.class);
            SecurityContextHolder.setContext(new SecurityContext(new Authentication(loginMember)));
        }

        return true;
    }

    @Override
    public void afterCompletion(HttpServletRequest request, HttpServletResponse response, Object handler, Exception ex) {
        SecurityContextHolder.clearContext();
    }
}
